server/requirements.txt
eventlet
python-socketio
pyautogui